backend/services/mental_health/app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
redis==5.0.1
# firebase-admin==6.2.0
google-generativeai==0.3.1
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
faiss-cpu==1.7.4
langchain==0.0.350
